LayoutTests/js/dom/modules/nomodule-has-no-effect-on-module-inline-expected.txt

Test nomodule does not have any effect on module script.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Module is not executed yet.
Module execution is confined in the module environment.
PASS typeof cocoa is "undefined"
PASS typeof exportedCocoa is "object"
PASS exportedCocoa.taste() is "awesome"
PASS successfullyParsed is true

TEST COMPLETE
